Add files via upload
This assignment focuses on sentiment analysis of IMDb reviews. We are provided with a large dataset, but only a small portion of it is labeled. We will utilize semi-supervised techniques to expand the labeled data and then train various models using this enhanced dataset. Additionally, we will explore the use of a Large Language Model (LLM) to achieve the same objective.